structural queries working in new refactor
