+++ get_traces_by_structure.h
@@ -4,23 +4,7 @@
 #include <iostream>                                                             
 #include <unordered_map>                                                        
 #include <utility>                                                              
-#include <map>                                                                  
-#include <string>                                                               
-#include <vector>                                                               
-#include <future>                                                               
-#include "google/cloud/storage/client.h"                                        
-#include "opentelemetry/proto/trace/v1/trace.pb.h"                              
-#include <boost/algorithm/string.hpp>                                           
-#include <boost/graph/adjacency_list.hpp>                                       
-#include <boost/graph/vf2_sub_graph_iso.hpp>                                    
-#include <boost/date_time/posix_time/posix_time.hpp>                            
 #include "common.h"
-
-const int element_count = 10000;
-
-typedef std::tuple<std::string, std::vector<std::string>> objname_to_matching_trace_ids;
-typedef std::tuple<std::string, std::vector<int>> trace_id_to_mappings;
-
 
 // TODO(jessica) trace struct and hashes should be keywords followed by SERVICES_BUCKETS_SUFFIX
 const char TRACE_STRUCT_BUCKET[] = "dyntraces-snicket4";
@@ -43,7 +27,9 @@ struct trace_structure {
     std::multimap<int, int> edges;
 };
 
-std::vector<std::string> split_by_char(std::string input, std::string splitter);
+// This is the highest level function
+std::vector<traces_by_structure> get_traces_by_structure(                       
+    trace_structure query_trace, int start_time, int end_time, gcs::Client* client);
 
 template < typename PropertyMapFirst, typename PropertyMapSecond >
 struct property_map_equivalent_custom {
@@ -70,6 +56,9 @@ struct property_map_equivalent_custom {
 };
 
 
+
+// Binary function object that returns true if the values for item1
+// in property_map1 and item2 in property_map2 are equivalent.
 template < typename PropertyMapFirst, typename PropertyMapSecond >
 property_map_equivalent_custom< PropertyMapFirst, PropertyMapSecond > make_property_map_equivalent_custom(
     const PropertyMapFirst property_map1, const PropertyMapSecond property_map2
@@ -85,14 +74,8 @@ graph_type;
 typedef boost::property_map<graph_type, boost::vertex_name_t>::type vertex_name_map_t;
 typedef property_map_equivalent_custom<vertex_name_map_t, vertex_name_map_t> vertex_comp_t;
 
-
-// Binary function object that returns true if the values for item1
-// in property_map1 and item2 in property_map2 are equivalent.
-
-
-
 /**
- * TODO: maybe pass the pointer to IsomorphismMaps, that should be a better practice.
+ * TODO(haseeb) maybe pass the pointer to IsomorphismMaps, that should be a better practice.
  */
 template < typename Graph1, typename Graph2, typename IsomorphismMaps>
 struct vf2_callback_custom {
@@ -122,18 +105,11 @@ struct vf2_callback_custom {
 
 std::vector<std::unordered_map<int, int>> get_isomorphism_mappings(
     trace_structure candidate_trace, trace_structure query_trace);
-std::map<std::string, std::string> get_trace_id_to_root_service_map(std::string object_content);
-std::map<std::string, std::vector<std::string>> get_root_service_to_trace_ids_map(
-    std::map<std::string, std::string> trace_id_to_root_service_map);
-std::vector<traces_by_structure> get_traces_by_structure(                       
-    trace_structure query_trace, int start_time, int end_time, gcs::Client* client);
 traces_by_structure process_trace_hashes_prefix_and_retrieve_relevant_trace_ids(
     std::string prefix, trace_structure query_trace, int start_time, int end_time,
     gcs::Client* client);
-std::string extract_batch_name(std::string object_name);
-std::pair<int, int> extract_batch_timestamps(std::string batch_name);
-std::string strip_from_the_end(std::string object, char stripper);
 trace_structure morph_trace_object_to_trace_structure(std::string trace);
 graph_type morph_trace_structure_to_boost_graph_type(trace_structure input_graph);
 std::vector<std::string> get_trace_ids_from_trace_hashes_object(std::string object_name, gcs::Client* client);
+void print_trace_structure(trace_structure trace);
 #endif // BY_STRUCT_H_
